+++ lib/rubocop/cop/style/redundant_parentheses.rb
@@ -23,20 +23,30 @@ module RuboCop
         def_node_matcher :range_end?, '^^{irange erange}'
         def_node_matcher :method_node_and_args, '$(send _recv _msg $...)'
         def_node_matcher :rescue?, '{^resbody ^^resbody}'
+        def_node_matcher :arg_in_call_with_block?,
+                         '^^(block (send _ _ equal?(%0) ...) ...)'
 
         def on_begin(node)
-          return unless parentheses?(node)
-
-          child_node = node.children.first
-          return if keyword_ancestor?(node) && parens_required?(node)
-          return if child_node.hash_type? && first_argument?(node) &&
-                    !parentheses?(node.parent)
-          return if rescue?(node)
-
-          check(node, child_node)
+          return if !parentheses?(node) || parens_allowed?(node)
+          check(node)
         end
 
-        def check(begin_node, node)
+        def parens_allowed?(node)
+          child  = node.children.first
+          parent = node.parent
+
+          # don't flag `break(1)`, etc
+          (keyword_ancestor?(node) && parens_required?(node)) ||
+            # don't flag `method ({key: value})`
+            (child.hash_type? && first_arg?(node) && !parentheses?(parent)) ||
+            # don't flag `rescue(ExceptionClass)`
+            rescue?(node) ||
+            # don't flag `method (arg) { }`
+            (arg_in_call_with_block?(node) && !parentheses?(parent))
+        end
+
+        def check(begin_node)
+          node = begin_node.children.first
           if keyword_with_redundant_parentheses?(node)
             return offense(begin_node, 'a keyword')
           end
@@ -60,7 +70,7 @@ module RuboCop
         end
 
         def keyword_ancestor?(node)
-          node.ancestors.first && node.ancestors.first.keyword?
+          node.parent && node.parent.keyword?
         end
 
         def disallowed_literal?(node)
@@ -90,7 +100,7 @@ module RuboCop
           args.empty? || parentheses?(send_node) || square_brackets?(send_node)
         end
 
-        def first_argument?(node)
+        def first_arg?(node)
           send_node = node.parent
           return false unless send_node && send_node.send_type?